requirements.txt
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
openai==1.7.2
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
python-dotenv==1.0.0
pypdf==3.17.4
pydantic==2.5.3
python-multipart==0.0.6
tiktoken==0.5.2
